crypto: Convert lengths in cipher API to unsigned int
- The cipher API used unsigned long for length parameters, in contrast
to the rest of the library which uses unsigned int.  This caused problems
when a length outparam stored by isrcry_cipher_final() was then passed to
e.g. isrcry_compress_final().  Convert cipher API to unsigned int length
parameters for consistency.

- Update callers

+++ crypto/cbc.c
@@ -42,7 +42,7 @@
   @return ISRCRY_OK if successful
 */
 static enum isrcry_result cbc_encrypt(struct isrcry_cipher_ctx *cctx,
-			const unsigned char *in, unsigned long len,
+			const unsigned char *in, unsigned len,
 			unsigned char *out)
 {
    unsigned blocklen = cctx->cipher->blocklen;
@@ -103,7 +103,7 @@ static enum isrcry_result cbc_encrypt(struct isrcry_cipher_ctx *cctx,
   @return ISRCRY_OK if successful
 */
 static enum isrcry_result cbc_decrypt(struct isrcry_cipher_ctx *cctx,
-			const unsigned char *in, unsigned long len,
+			const unsigned char *in, unsigned len,
 			unsigned char *out)
 {
    unsigned blocklen = cctx->cipher->blocklen;
